Initial version of Chandra.
(decided to name it to difftiate it from the other two)

At present, it sucessfully simulates decks with only burn cards
and uses a genetic algorithm to optimize for dealing 20 damage
as quickly as possible.

The best deck it's gotten so far in the time I've run it for is
8 turns (average) using only cards in standard, and 6.4 turns with
all the cards it knows about. The standard deck appears to be completely
converged because it includes as many as possible of all but one of the cards
it's using and no others, while the all-inclusive deck is close but not
completely optimized.