Started on my genetic algorithm
Added one type of mutation, crossover, initial population and next
population generation. Want to add different types later. Still need to
combine the parts and add a function to run the population through the
simulation. Added a function in Player to generate a blank Player for
various instances where I use a card but don't care about the player of
the card.